Implement redux/elm-like architecture
The `store` type encapsulates the mutable variable that holds the
state.  It serializes updates via the `dispatch()` method, that is
thread safe.  It propagates the data across updates.

This system system is most similar to Elm, because the reducer can
actually optionally return an effect.  The idea is that now update()
can actually really be a pure function---before, stuff like the "save"
command would execute side-effects immediatelly.  When an update()
would trigger an effect, it now returns a function that will be later
evaluated by the store, passing in the required information so it can
safely start effect chains that may in the future produce more
actions.